+++ controllers/base.js
@@ -44,7 +44,7 @@ module.exports.deleteExercise = async function (req, res) {
         const base = await Base.findById(req.params.id);
 
         if (base.use > 0) {
-            return res.status(403).json({message: 'You cannot delete this exercise because it is in use!'});
+            return res.status(403).json({ message: 'You cannot delete this exercise because it is in use!' });
         }
 
         fs.unlink(path.resolve(serverPath.path, base.video), async (err) => {
@@ -69,7 +69,41 @@ module.exports.updateUse = async function (req, res) {
             { new: true }
         );
 
-        res.status(200).json({message: 'Updated!'});
+        res.status(200).json({ message: 'Updated!' });
+    } catch (e) {
+        errorHandler(res, e);
+    }
+}
+
+module.exports.update = async function (req, res) {
+    try {
+        const compare = await Base.findById(req.params.id);
+
+        if (req.file) {
+            fs.unlink(path.resolve(serverPath.path, compare.video), async (err) => {
+                if (err) {
+                    return res.status(500).json({ message: 'Failed to update exercise!' });
+                }
+                else {
+                    const base = await Base.findByIdAndUpdate(
+                        { _id: req.params.id },
+                        { $set: { name: req.body.name, video: req.file.path } },
+                        { new: true }
+                    );
+
+                    return res.status(200).json({ message: 'Updated!' });
+                }
+            });
+        }
+        else {
+            const base = await Base.findByIdAndUpdate(
+                { _id: req.params.id },
+                { $set: { name: req.body.name } },
+                { new: true }
+            );
+
+            res.status(200).json({ message: 'Updated!' });
+        }
     } catch (e) {
         errorHandler(res, e);
     }